+++ src/Pim/Bundle/CatalogBundle/Doctrine/ORM/Filter/EntityFilter.php
@@ -4,7 +4,6 @@ namespace Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter;
 
 use Doctrine\ORM\QueryBuilder;
 use Pim\Bundle\CatalogBundle\Doctrine\Query\FieldFilterInterface;
-use Pim\Bundle\CatalogBundle\Context\CatalogContext;
 
 /**
  * Entity filter
@@ -18,9 +17,6 @@ class EntityFilter implements FieldFilterInterface
     /** @var QueryBuilder */
     protected $qb;
 
-    /** @var CatalogContext */
-    protected $context;
-
     /** @var array */
     protected $supportedFields;
 
@@ -30,12 +26,10 @@ class EntityFilter implements FieldFilterInterface
     /**
      * Instanciate the base filter
      *
-     * @param CatalogContext $context
-     * @param array          $extraSupportedFields
+     * @param array $extraSupportedFields
      */
-    public function __construct(CatalogContext $context, array $extraSupportedFields = [])
+    public function __construct(array $extraSupportedFields = [])
     {
-        $this->context = $context;
         $this->supportedFields = array_merge(
             ['family', 'groups'],
             $extraSupportedFields