+++ src/Pim/Bundle/CatalogBundle/Doctrine/ORM/Sorter/CompletenessSorter.php
@@ -5,7 +5,6 @@ namespace Pim\Bundle\CatalogBundle\Doctrine\ORM\Sorter;
 use Doctrine\ORM\QueryBuilder;
 use Pim\Bundle\CatalogBundle\Doctrine\Query\FieldSorterInterface;
 use Pim\Bundle\CatalogBundle\Doctrine\ORM\CompletenessJoin;
-use Pim\Bundle\CatalogBundle\Context\CatalogContext;
 
 /**
  * Completeness sorter
@@ -16,24 +15,8 @@ use Pim\Bundle\CatalogBundle\Context\CatalogContext;
  */
 class CompletenessSorter implements FieldSorterInterface
 {
-    /**
-     * QueryBuilder
-     * @var QueryBuilder
-     */
+    /** @var QueryBuilder */
     protected $qb;
-
-    /** @var CatalogContext */
-    protected $context;
-
-    /**
-     * Instanciate a sorter
-     *
-     * @param CatalogContext $context
-     */
-    public function __construct(CatalogContext $context)
-    {
-        $this->context = $context;
-    }
 
     /**
      * {@inheritdoc}